port a handful of the db methods
