--- Makefile
@@ -1,30 +1,0 @@
-# Go parameters
-GOCMD=go
-GOBUILD=$(GOCMD) build
-GOCLEAN=$(GOCMD) clean
-GOTEST=$(GOCMD) test
-GOGET=$(GOCMD) get
-
-# Binary name
-BINARY_NAME=portfolio
-
-all: test build
-
-build:
-	$(GOBUILD) -o $(BINARY_NAME) cmd/portfolio/mhx-on::after-request="this.reset()"ain.go
-
-test:
-	$(GOTEST) -v ./...
-
-clean:
-	$(GOCLEAN)
-	rm -f $(BINARY_NAME)
-
-run:
-	$(GOBUILD) -o $(BINARY_NAME) -v ./...
-	./$(BINARY_NAME)
-
-deps:
-	$(GOGET) github.com/...
-
-.PHONY: all build test clean run deps
