+++ go.mod
@@ -4,6 +4,7 @@ go 1.20
 
 require (
 	github.com/joho/godotenv v1.5.1
+	github.com/mattn/go-sqlite3 v1.14.17
 	go.mongodb.org/mongo-driver v1.12.1
 )
