+++ templates/link.html
@@ -1,25 +1,24 @@
 {{define "link"}}
-<div class="list-group-item list-group-item-action" id="link-{{ oidToStr .Id }}">
+<div class="list-group-item list-group-item-action" id="link-{{ idToStr .Id }}">
   <div class="d-flex w-100 justify-content-between">
     <div>
       <h5 class="mb-1"><a href="{{ .Url }}">{{ .Name }}</a></h5>
-      <p class="mb-1">{{ range .Labels }}<span class="badge bg-secondary me-2">{{ .Name }}</span>{{ end }}</p>
       <small>{{ .Url }}</small>
     </div>
     <div class="align-self-center">
       <div>
         <button
           class="btn btn-primary"
-          hx-get="/link/edit?id={{ oidToStr .Id }}"
-          hx-target="#link-{{ oidToStr .Id }}"
+          hx-get="/link/edit?id={{ idToStr .Id }}"
+          hx-target="#link-{{ idToStr .Id }}"
           hx-swap="innerHTML"
         >
           Edit
         </button>
         <button
           class="btn btn-danger ms-2"
-          hx-delete="/link/delete?id={{ oidToStr .Id }}"
-          hx-target="#link-{{ oidToStr .Id }}"
+          hx-delete="/link/delete?id={{ idToStr .Id }}"
+          hx-target="#link-{{ idToStr .Id }}"
           hx-swap="outerHTML"
         >
           Delete
